front end

+++ cart.php
@@ -1,10 +1,24 @@
 <?php
 include('admin/config.php');
-if (isset($_GET['id'])) {
+
+if (isset($_GET['id']) && isset($_GET['action'])) {
 
   if ($_GET['action'] == 'remove') {
     $id = $_GET['id'];
     $sql = "DELETE from cart WHERE id='$id'";
+    $result = $conn->query($sql);
+  }
+}
+
+$x = 0;
+if (isset($_GET['update'])) {
+
+  $quant = $_GET['quant'];
+
+  foreach ($quant as $key => $val) {
+    $id = $key;
+    $quantity = $val;
+    $sql = "UPDATE cart SET quantity =$quantity WHERE id = $id;";
     $result = $conn->query($sql);
   }
 }
@@ -35,7 +49,7 @@ if (isset($_GET['id'])) {
       <div class="col-md-12">
         <div class="cart-view-area">
           <div class="cart-view-table">
-            <form action="">
+            <form action="" method="GET">
               <div class="table-responsive">
                 <table class="table">
                   <thead>
@@ -49,10 +63,12 @@ if (isset($_GET['id'])) {
                     </tr>
                   </thead>
                   <tbody>
+
                     <?php
                     $totalprice = 0;
                     $sql = "SELECT * FROM cart ";
                     $result = $conn->query($sql);
+
                     if ($result->num_rows > 0) {
                       while ($row = $result->fetch_assoc()) {
                     ?>
@@ -64,20 +80,22 @@ if (isset($_GET['id'])) {
                           <td><a href="#"><img src="admin/resources/pimages/<?php echo $row['image']; ?>" alt="img"></a></td>
                           <td><a class="aa-cart-title" href="#"><?php echo $row['name']; ?></a></td>
                           <td>$<?php echo $row['price']; ?></td>
-                          <td><input class="aa-cart-quantity" type="number" value="<?php echo $row['quantity']; ?>"></td>
+                          <td><input class="aa-cart-quantity" type="number" name="quant[<?php echo $row['id']; ?>]" value="<?php echo $row['quantity']; ?>"></td>
                           <td>$<?php $total = $row['price'] * $row['quantity'];
                                 $totalprice += $total;
                                 echo $total; ?></td>
                         </tr>
+
                       <?php } ?>
                     <?php } ?>
+
                     <tr>
                       <td colspan="6" class="aa-cart-view-bottom">
                         <div class="aa-cart-coupon">
                           <input class="aa-coupon-code" type="text" placeholder="Coupon">
                           <input class="aa-cart-view-btn" type="submit" value="Apply Coupon">
                         </div>
-                        <input class="aa-cart-view-btn" type="submit" value="Update Cart">
+                        <input class="aa-cart-view-btn" type="submit" name="update" value="update" />
                       </td>
                     </tr>
                   </tbody>
